+++ include/game/player.h
@@ -16,7 +16,7 @@ typedef struct
 typedef struct
 {
     float speed_multiplier;
-    int power;
+    float power;
     float max_speed;
 } BallAttributes;
